GEODE-7503: Cleanup GemFireCacheImpl (#4436)
* Move inner-classes to bottom of outer class

* Move fields to top of class

* More fields

* Fixup

* Fix IJ inspection warnings

* Remove bug system ticket numbers

* Cleanup javadocs, toString, and comments

* Use @VisibleForTesting annotation

* Move toString method and backupFiles field

* Pull javadocs up to interfaces.
Define VisibleForTesting methods in InternalCacheForTesting.
Identify more methods as VisibleForTesting.

* Fixup javadocs and minor warnings in Cache interfaces.

* Move unchecked util methods to UncheckedUtils inner class

* Move static constants and fields to top of class.
Rename FIVE_HOURS to FIVE_HOURS_MILLIS.
Review older TODOs.

* Move volatile fields above non-final fields

* Move non-final fields below volatile and final fields

* Move final fields above volatile fields

* wip

* Adjust whitespace and line feeds.
Remove more TODOs and comments.
Fix spelling typo.

* Import various static methods and constants
Adjust more line feeds
Adjust more white space

* Weaken declared types where it makes sense
Add more static imports

* Remove unnecessary uses of final on local vars

* Remove uncessary uses of final on local vars

* Fixup unchecked types
Restore return types to match the interfaces

* Minor cleanup

* Change Consumer<Void> to Runnable (duh)

* Adjust line feeds of large if-statement

* Remove SuppressWarnings

* Fix AssemblyContentsIntegrationTest

* Fixup GemFireCacheImpl

+++ geode-core/src/main/java/org/apache/geode/internal/admin/remote/RemoteCacheInfo.java
@@ -62,7 +62,7 @@ public class RemoteCacheInfo implements CacheInfo, DataSerializable {
     this.lockTimeout = internalCache.getLockTimeout();
     this.lockLease = internalCache.getLockLease();
     this.searchTimeout = internalCache.getSearchTimeout();
-    this.upTime = internalCache.getUpTime();
+    this.upTime = (int) internalCache.getUpTime();
     if (this.closed) {
       this.rootRegionNames = null;
       this.perfStats = null;
